// src/flight-alerts/domain/enums/flight-alert-type.enum.ts
export enum FlightAlertType {
  PRICE_DROP = 'PRICE_DROP',
  PRICE_DROP_CHEAPEST = 'PRICE_DROP_CHEAPEST',
  PRICE_DROP_RECOMMENDED = 'PRICE_DROP_RECOMMENDED',
  TARGET_PRICE_REACHED = 'TARGET_PRICE_REACHED',
  LOWEST_HISTORICAL_PRICE = 'LOWEST_HISTORICAL_PRICE',
  DAILY_SUMMARY = 'DAILY_SUMMARY',
}